requirements.txt
psycopg2~=2.9.10
fastapi~=0.116.1
pydantic~=2.11.7
SQLAlchemy~=2.0.41
uvicorn~=0.35.0
passlib~=1.7.4
bcrypt~=4.0.1
email-validator~=2.2.0
python-jose~=3.5.0
cryptography~=45.0.5